Move the `app-setup` module under the `@bayou` namespace.
